Clean code v5 - Library feature

+++ lib/features/library/data/repositories/library_repository_impl.dart
@@ -11,7 +11,7 @@ class LibraryRepositoryImpl implements LibraryRepository {
   Future<void> addStoryToBookmarks(String storyId) async {
     final user = _supabaseClient.auth.currentUser;
     if (user == null) throw const AuthException('User is not authenticated');
-    
+
     await _supabaseClient.from('User_Bookmarked_Stories').insert({
       'user_id': user.id,
       'story_id': storyId,
@@ -23,12 +23,13 @@ class LibraryRepositoryImpl implements LibraryRepository {
     final user = _supabaseClient.auth.currentUser;
     if (user == null) return [];
 
+    // Select bookmarked stories and join with the Story and Profile tables.
     final data = await _supabaseClient
         .from('User_Bookmarked_Stories')
         .select('*, Story!inner(*, profiles:author_id(*))')
         .eq('user_id', user.id)
         .order('bookmarked_at', ascending: false);
-    
+
     return data.map((item) => Story.fromMap(item['Story'])).toList();
   }
 
@@ -53,10 +54,6 @@ class LibraryRepositoryImpl implements LibraryRepository {
     final user = _supabaseClient.auth.currentUser;
     if (user == null) throw const AuthException('User is not authenticated');
 
-    await _supabaseClient
-        .from('User_Bookmarked_Stories')
-        .delete()
-        .eq('user_id', user.id)
-        .eq('story_id', storyId);
+    await _supabaseClient.from('User_Bookmarked_Stories').delete().eq('user_id', user.id).eq('story_id', storyId);
   }
 }
